Centralize sqlite connection and wrapping code
