+++ stepup/core/utils.py
@@ -20,7 +20,6 @@
 """Small utilities used throughout."""
 
 import asyncio
-import contextlib
 import logging
 import os
 import re
@@ -28,7 +27,7 @@ import shlex
 import sqlite3
 import string
 import sys
-from collections.abc import Collection, Iterator
+from collections.abc import Collection
 
 import attrs
 from path import Path
@@ -429,18 +428,3 @@ def string_to_bool(v: str | bool) -> bool:
             return False
         raise ValueError(f"Cannot interpret '{v}' as a boolean value.")
     raise TypeError(f"Expected a boolean value or string. Got {type(v).__name__}")
-
-
-@contextlib.contextmanager
-def sqlite3_copy_in_memory(path_db) -> Iterator[sqlite3.Connection]:
-    """Copy an SQLite database into memory and yield the connection."""
-    dst = sqlite3.Connection(":memory:", detect_types=sqlite3.PARSE_COLNAMES)
-    try:
-        src = sqlite3.Connection(path_db, detect_types=sqlite3.PARSE_COLNAMES)
-        try:
-            src.backup(dst)
-        finally:
-            src.close()
-        yield dst
-    finally:
-        dst.close()
